Move most of storage stuff to storage package
