+++ rawclient.go
@@ -4,6 +4,8 @@ import (
 	"errors"
 	"net/http"
 	"net/url"
+
+	"github.com/grafana/nanogit/storage"
 )
 
 type rawClient struct {
@@ -20,7 +22,7 @@ type rawClient struct {
 	// Token-based authentication header
 	tokenAuth *string
 	// Packfile storage
-	packfileStorage PackfileStorage
+	packfileStorage storage.PackfileStorage
 }
 
 // addDefaultHeaders adds the default headers to the request.
