+++ tree.go
@@ -7,6 +7,7 @@ import (
 
 	"github.com/grafana/nanogit/protocol"
 	"github.com/grafana/nanogit/protocol/hash"
+	"github.com/grafana/nanogit/storage"
 )
 
 // FlatTreeEntry represents a single entry in a flattened Git tree structure.
@@ -110,7 +111,7 @@ func (c *httpClient) GetFlatTree(ctx context.Context, commitHash hash.Hash) (*Fl
 
 // fetchAllTreeObjects collects all tree objects needed for the flat tree by starting with
 // an initial request and iteratively fetching missing tree objects in batches.
-func (c *httpClient) fetchAllTreeObjects(ctx context.Context, commitHash hash.Hash) (PackfileStorage, hash.Hash, error) {
+func (c *httpClient) fetchAllTreeObjects(ctx context.Context, commitHash hash.Hash) (storage.PackfileStorage, hash.Hash, error) {
 	logger := c.getLogger(ctx)
 	// Track essential metrics
 	var totalRequests int
@@ -336,7 +337,7 @@ func (c *httpClient) fetchAllTreeObjects(ctx context.Context, commitHash hash.Ha
 // collectMissingTreeHashes processes tree objects and collects missing child tree hashes.
 // It iterates through the provided objects, optionally adds them to allObjects if addToCollection is true,
 // and identifies any missing child tree objects that need to be fetched.
-func (c *httpClient) collectMissingTreeHashes(ctx context.Context, objects map[string]*protocol.PackfileObject, allObjects PackfileStorage, pending []hash.Hash, processedTrees map[string]bool, requestedHashes map[string]bool) ([]hash.Hash, error) {
+func (c *httpClient) collectMissingTreeHashes(ctx context.Context, objects map[string]*protocol.PackfileObject, allObjects storage.PackfileStorage, pending []hash.Hash, processedTrees map[string]bool, requestedHashes map[string]bool) ([]hash.Hash, error) {
 	logger := c.getLogger(ctx)
 	var treesProcessed int
 	var newTreesFound int
@@ -416,7 +417,7 @@ func (c *httpClient) collectMissingTreeHashes(ctx context.Context, objects map[s
 
 // findRootTree locates the root tree object from the target hash and available objects.
 // It handles both commit and tree target objects, extracting the tree hash and object as needed.
-func (c *httpClient) findRootTree(ctx context.Context, targetHash hash.Hash, allObjects PackfileStorage) (*protocol.PackfileObject, hash.Hash, error) {
+func (c *httpClient) findRootTree(ctx context.Context, targetHash hash.Hash, allObjects storage.PackfileStorage) (*protocol.PackfileObject, hash.Hash, error) {
 	logger := c.getLogger(ctx)
 	// Find our target object in the response
 	obj, exists := allObjects.Get(targetHash)
@@ -455,7 +456,7 @@ func (c *httpClient) findRootTree(ctx context.Context, targetHash hash.Hash, all
 }
 
 // flatten converts collected tree objects into a flat tree structure using breadth-first traversal.
-func (c *httpClient) flatten(treeHash hash.Hash, allTreeObjects PackfileStorage) (*FlatTree, error) {
+func (c *httpClient) flatten(treeHash hash.Hash, allTreeObjects storage.PackfileStorage) (*FlatTree, error) {
 	// Get the root tree object
 	rootTree, exists := allTreeObjects.Get(treeHash)
 	if !exists {
