+++ writer.go
@@ -9,6 +9,7 @@ import (
 
 	"github.com/grafana/nanogit/protocol"
 	"github.com/grafana/nanogit/protocol/hash"
+	"github.com/grafana/nanogit/storage"
 )
 
 // NewStagedWriter creates a new StagedWriter for staging changes to a Git reference.
@@ -99,7 +100,7 @@ type stagedWriter struct {
 	// Root tree object from last commit
 	lastTree *protocol.PackfileObject
 	// Cache of fetched tree objects
-	objStorage PackfileStorage
+	objStorage storage.PackfileStorage
 	// Flat mapping of paths to tree entries
 	treeEntries map[string]*FlatTreeEntry
 }
